third_party/WebKit/LayoutTests/media/track/track-default-attribute.html

<!DOCTYPE html>
<title>Tests that a track with the "default" attribute loads automatically.</title>
<video>
    <track kind="captions" src="captions-webvtt/tc005-default-styles.vtt">
    <track kind="captions" src="captions-webvtt/tc005-metadata-area.vtt">
    <track default kind="captions" src="captions-webvtt/tc004-webvtt-file.vtt" id="default">
</video>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var timer = null;
    var tracks = document.querySelectorAll("track");
    for (var track of tracks) {
        track.onload = t.step_func(function() {
            assert_equals(event.target.readyState, HTMLTrackElement.LOADED);
            assert_equals(event.target.id, "default");
            assert_true(event.target.default);
            // End the test after a brief pause so we allow other tracks to load if they will.
            if (timer)
                clearTimeout(timer);
            timer = setTimeout(function() { t.done(); }, 200);
        });
    }
});
</script>
